Migrate data-table from /docs/components/ to /components/
Move Filtering and Row Selection examples from the old documentation
page to the new reference page, delete the old page, remove its route,
and consolidate E2E tests to target /components/data-table.
